+++ R/config.R
@@ -199,6 +199,23 @@ py_discover_config <- function(required_module = NULL, use_environment = NULL) {
     
   }
   
+  # if we're working within a project that contains a pyproject.toml file,
+  # then use the copy of Python associated with the poetry environment
+  projfile <- poetry_project_path()
+  if (file.exists(projfile)) {
+    
+    # validate that 'poetry' is available
+    poetry <- poetry_binary_path()
+    if (!file.exists(poetry)) {
+      warning("This project contains a 'pyproject.toml' file, but 'poetry' is not available")
+    } else {
+      python <- poetry_python_path(dirname(projfile))
+      config <- python_config(python, required_module, forced = "Poetry")
+      return(config)
+    }
+    
+  }
+  
   # if we're working within a project that contains a Pipfile, then
   # use the copy of Python associated with that pipenv
   pipfile <- pipenv_pipfile_path()